added check for exhausted memory
if many emails will be synchronised, we can easily exhaust PHP memory.
This change add a the new Syncroton_Exception_MemoryExchausted which can
be thrown in the backend, if the backend decides that the current entry
will exhaust PHP memory.
If the MemoryExchausted exception will be thrown, the entry will be
skipped and will be synchronised during the next sync.

Change-Id: I3fb82830c7c58f96231388d1b5b6a85d3a7aebc1