Speedup of Record::__get function
The __get function does not need to check validators as this is checked
upon __set.
So logically the value is either set or not and does not need to be
validated,
other than trowing an Exception for debugging purposes.
Should be stable by now... :-)
This modification will speed this VERY OFTEN called function up by
factor 2

Change-Id: Ib89b7f929f9e48b1d1854cea8babc3af5f15a7d0
Reviewed-on: https://gerrit.tine20.org/tine20/2664
Tested-by: jenkins user
Reviewed-by: Lars Kneschke <l.kneschke@metaways.de>