unnecessary query
tine20/Tinebase/Backend/Sql/Abstract.php - search() selects ids first and then eventually makes a second query to fetch all the columns. Added logic to determin wether there is a difference in the joins of the first query vs. the second query. If first and second query are using the same joins, there is no sense in doing both queries. In that case now only one query will be performed. If there is a difference in the joins, then the logic as is will be executed.

Change-Id: Ib93584661ee1021c868a189603c79a50d16f3db5
Reviewed-on: https://gerrit.tine20.org/tine20/2627
Tested-by: jenkins user
Reviewed-by: Philipp Schüle <p.schuele@metaways.de>