0009994: Use css classes instead of direct formats
Hi, I noticed that the Background Colors for the Day cells in the 
Monthview Calendar, were set directly through a javascript file. 
So it is not possible to change the colors in theme.

a detailed description can be found in the german sub-forum:
https://www.tine20.org/forum/viewtopic.php?f=12&t=16272#p57288

and a possible solution too...
https://www.tine20.org/forum/viewtopic.php?f=12&t=16272#p57314

(by ravermeister)

https://forge.tine20.org/mantisbt/view.php?id=9994

Change-Id: I00684e08f3b5909d4b45f22985614f01b694ed5e
Reviewed-on: https://gerrit.tine20.org/tine20/2846
Reviewed-by: Philipp Schüle <p.schuele@metaways.de>
Tested-by: Philipp Schüle <p.schuele@metaways.de>