Invoice deletion now based on the actual invoice position
Sales_Controller_Invoice::_inspectDelete():
The product aggregates last_autobill is set back by the actually existing
(and now to be deleted) invoice positions. That is a real undo.
Before the reset of the last_autobill was based on speculation and thus error prone.

Invoice creation now iterates over the months from the first date that needs to be
billed until the current billing month and creates multiple invoices in one go as needed.

Change-Id: I9481d6ad603e18aef92aec076f3ea73ad471f303
Reviewed-on: http://gerrit.tine20.com/customers/1718
Reviewed-by: Philipp Schüle <p.schuele@metaways.de>
Tested-by: Jenkins CI (http://ci.tine20.com/)