VCard converter for gnome Evolution client >= 3.10.4
The Gnome Email and Groupware client named Evolution is currently not
supported by Tine20. Therefore new records could be added via generic
CardDav support but updating existing records was refuesed cause of missing
convertter for Evolution client.

Adds converter for importing Gnome Evolution CardDav data into tine20 data
model and export tine20 Addressbook to CardDav format understood by Gnome
Evolution. Currently tested with Gnome Evolution 3.10.4 and 10.18.5.2

Moved logging message in default case of switch in _toTine20Model function
of Addressbook_Convert_Contact_VCard_Abstract to _toTine20ModelParseOther
This function is overriden by evolution client converter to parse the
X-Evolution-<Type> fields and assign them to their corresponding tine20
model item.

Extended PHOTO case in _toTine20Model within PHP to safely handle Photos
Which for example are allways assigned to X-Evolution-Unknown Type instead
of their proper type JPEG or PNG. The chosen solution check the TYPE tag. If
it reads JPEG or PNG as expected it just passes the image on as before.
In any other case it tries to identify the mimetype of the image and
if not supported by tine20 it tries to convert it to jpeg. If this is
not possible or the encoding of the image is not binary the content is
passed on as is without any changes.

Adds support for parsing and sending room information. Not fully
supported by Tine20 yet.

_fromTine20ModelRequiredFields now takes two additional parameters $fn and
$org allowing the converter to set custom values for the $fn and $org field.
If null or note set $fn = $record->n_fileas and
$org = array($record->org_name, $record->org_unit);
Both are used by Evolution converter to add room to ORG field and add custom
X_EVOLUTION_FILEAS field and recoding FN to full name string.

Partial-Bug: #0011006