Support for Gnome Evolution tasks
In current version of Tine20 it is not possible to add or update task
using the Gnome Evolution client.

Adds converter for Gnome Evolution client, which was tested on Gnome
Evolution 3.10.04.

For new tasks by Gnome Evoultution (at least 3.10.04 and 3.18.5.2) uses the following
patterns to generated uid and
the url of the .ics file.
   -> uid : <creattiontimestamp><identifier>@<hostname>
   -> .ics: <uid>-<modificationtimestamp>

<creationtimestpamp>: <timestamp>
<modificationtimestamp>: <timestamp>
<timestamp>: <YYYY><MM><DD>T<hh><mm><ss>Z
<YYYY>: four digit year number
<MM>: two digit month number
<DD>: two digit month day number
<hh>: two digit hour
<mm>: two digit minutes
<ss>: two digit seconds
<hostname>: hostname as returned by `hostname` command

The resulting record id is longer than 40 char, and it will be in many
cases be unique causing multiple instances of the same record be created,
as urls will differ by <modificationtimestamp>.

In order to get consistent ids the converter checks the uid if it contains
the host name it strips off the host name and removes any '_' and '-'
characters from the uid string and explicitly sets the id of the task
record to this newly created id string. The
Tasks_Frontend_WebDAV_Task::createFile function checks whether the
converter has set an id for the task record if it is set than it does not
use the url of the .ics file to set the id of the task record.

Related-Bug: #0011010