Changes required for proper https only access to remote dev-system
Master branch is setup to be used inhouse where the development server
and the testing client are within the same secured network.

These fixes are merely for my setup where the testing and development
server is located in a remote location which requires to use https for
accessing all non static html based services via https.

It further contains a small tweak for my my setup using webpack 2.5.1
instead of 1.13.3