Tinebase Export - allow twig extension registering
used in tests for "cache" busting (once a php class got generated and
exists in the current process, it is not unloaded anymore. If the same
hash used as class name is generated again, the class exists already
and will be reused. Avoided by loading an extension with random name)

Change-Id: I5223e3a94469dca1df559ef60c35033e3c29fc1b
Reviewed-on: http://gerrit.tine20.com/customers/8329
Tested-by: Jenkins CI (http://ci.tine20.com/) <tine20-jenkins@metaways.de>
Reviewed-by: Paul Mehrer <p.mehrer@metaways.de>