tweak(Tinebase Core) - try to create tmp dir if it doesn't exists
before falling back to default locations, try to create the tmp dir
in case it doesn't exists.

Also create a instance unique hash based on the configured tmp dir,
in case we can't create it. Then append that to the fall back locations
so that in a multi instance environment, multiple instance can fall
back to the default location and still have each separate tmp dirs.

Change-Id: I726333359b0ae5d44461dc1d1edca45e020504fd
Reviewed-on: http://gerrit.tine20.com/customers/10569
Tested-by: Jenkins CI (http://ci.tine20.com/) <tine20-jenkins@metaways.de>
Reviewed-by: Paul Mehrer <p.mehrer@metaways.de>