Commit 2f083b0 breaks Ldap login
The function _getConfiguredSystemDefaults() was moved to Tinebase/User/Plugin/Abstract which serves SQL-based logins only. Before it was in Tinebase/EmailUser/Smtp/Standard.php, from where all Plugins are inherited. 

This is meant as a hot fix. In principle there should be a common place for this regardless sql/ldap account storage (needs refactoring).