feature(ci) reproducible and updateable docker builds
Docker images are build from cache to archive reproducibility.
In case all layers are pulled from cache, we use use the cache
image. This way the image sha dose not change. There are
"cache-invalidator" images. They are build with no cache and
produce a singel file. The file should change if and only if
the cache needs to be invalidatet. E.g. apk add name install a
newer version on name updates packets, then the file should
change. "provider" images are a basicly the same as a cache-
invalidator, but they provide multible files. E.g. a provider
provieds the icon-set. The icon-set is a git submodule. We
can not copy the .git folder into the main image. The provider
initilies the submodule and the main image copys the icon-set
folder.

Change-Id: Ib412c7a6dfa1eff3c5ce9d3e77f4fb7617782ba2
Reviewed-on: http://gerrit.tine20.com/customers/18065
Reviewed-by: Milan Mertens <m.mertens@metaways.de>
Tested-by: Milan Mertens <m.mertens@metaways.de>