fix(Calendar/CalDAV): fix URI of deleted events
This commit fixes the URI of deleted calendar events in REPORT responses.

The URIs for deleted events are handled in [1] directly, while the URIs for
all other events (created, changed, etc.) are handled in [2].

Both code paths should result in the exact same URIs. At least the
Thunderbird Lightning extension relies on the URIs to match.

[1] Tinebase_WebDav_Plugin_SyncToken::generateMultiStatus
[2] Sabre\DAV\Property\Response::serialize